+++ Game/chest.h
@@ -10,7 +10,7 @@ class chest : public sf::Sprite
 public:
     chest();
     int room_number();
-    void room_number_set(int number);
+    void room_number_set(unsigned long long number);
     void open();
     bool is_oppened();
     Item itemik;
@@ -18,7 +18,7 @@ private:
     sf::Texture Texture1,Texture2;
     int rarity_;
     bool clossed_;
-    int room_number_;
+    unsigned long long room_number_;
 };
 
 #endif // CHEST_H
